tracer: restructure PrivateTraceAPI to its own file
